need to check if the file exists before we try to remove it

+++ system/database/drivers/pdo/pdo_driver.php
@@ -77,7 +77,8 @@ class CI_DB_pdo_driver extends CI_DB
      */
     function db_pconnect()
     {
-        unlink("sof.db-journal");
+        if (file_exists("sof.db-journal"))
+            unlink("sof.db-journal");
         try
         {
             $conn_id = new PDO ($this->database, $this->username, $this->password, array(PDO::ATTR_PERSISTENT => true));
